refactor and fix create wallet bug
