+++ src/components/lns/lns_record_list.vue
