+++ src/components/menus/wallet_settings.vue
